Add files via upload
Extract potential features of protein sequence to identify the types of proteins. Design novel feature selection framework or use the existing techniques with proper justifications to identify salient features of the given data. Report the performance of different classification techniques on the training data to demonstrate that the proposed feature selection scheme is working well. You can also propose a novel classification framework for this problem. Subsequently, execute the best framework on the test data and submit the class labels in a text file. Each row of the text file will contain the class label of an instance of the test data e.g., Hydrolase following the order of the given test data.